Move some instructions methods next to where they are used
