+++ compiler/noirc_evaluator/src/ssa/ir/instruction.rs
@@ -17,7 +17,6 @@ use super::{
     basic_block::BasicBlockId,
     call_stack::CallStackId,
     dfg::DataFlowGraph,
-    function::Function,
     map::Id,
     types::{NumericType, Type},
     value::{Value, ValueId},
@@ -395,267 +394,6 @@ impl Instruction {
     /// inserting this instruction into a DataFlowGraph.
     pub(crate) fn requires_ctrl_typevars(&self) -> bool {
         matches!(self.result_type(), InstructionResultType::Unknown)
-    }
-
-    /// Indicates if the instruction has a side effect, ie. it can fail, or it interacts with memory.
-    ///
-    /// This is similar to `can_be_deduplicated`, but it doesn't depend on whether the caller takes
-    /// constraints into account, because it might not use it to isolate the side effects across branches.
-    pub(crate) fn has_side_effects(&self, dfg: &DataFlowGraph) -> bool {
-        use Instruction::*;
-
-        match self {
-            // These either have side-effects or interact with memory
-            EnableSideEffectsIf { .. }
-            | Allocate
-            | Load { .. }
-            | Store { .. }
-            | IncrementRc { .. }
-            | DecrementRc { .. } => true,
-
-            Call { func, .. } => match dfg[*func] {
-                Value::Intrinsic(intrinsic) => intrinsic.has_side_effects(),
-                // Functions known to be pure have no side effects.
-                // `PureWithPredicates` functions may still have side effects.
-                Value::Function(function) => dfg.purity_of(function) != Some(Purity::Pure),
-                _ => true, // Be conservative and assume other functions can have side effects.
-            },
-
-            // These can fail.
-            Constrain(..) | ConstrainNotEqual(..) | RangeCheck { .. } => true,
-
-            // This should never be side-effectful
-            MakeArray { .. } | Noop => false,
-
-            // Some binary math can overflow or underflow
-            Binary(binary) => match binary.operator {
-                BinaryOp::Add { unchecked: false }
-                | BinaryOp::Sub { unchecked: false }
-                | BinaryOp::Mul { unchecked: false }
-                | BinaryOp::Div
-                | BinaryOp::Mod => true,
-                BinaryOp::Add { unchecked: true }
-                | BinaryOp::Sub { unchecked: true }
-                | BinaryOp::Mul { unchecked: true }
-                | BinaryOp::Eq
-                | BinaryOp::Lt
-                | BinaryOp::And
-                | BinaryOp::Or
-                | BinaryOp::Xor
-                | BinaryOp::Shl
-                | BinaryOp::Shr => false,
-            },
-
-            // These can have different behavior depending on the EnableSideEffectsIf context.
-            Cast(_, _)
-            | Not(_)
-            | Truncate { .. }
-            | IfElse { .. }
-            | ArrayGet { .. }
-            | ArraySet { .. } => self.requires_acir_gen_predicate(dfg),
-        }
-    }
-
-    /// Indicates if the instruction can be safely replaced with the results of another instruction with the same inputs.
-    /// If `deduplicate_with_predicate` is set, we assume we're deduplicating with the instruction
-    /// and its predicate, rather than just the instruction. Setting this means instructions that
-    /// rely on predicates can be deduplicated as well.
-    ///
-    /// Some instructions get the predicate attached to their inputs by `handle_instruction_side_effects` in `flatten_cfg`.
-    /// These can be deduplicated because they implicitly depend on the predicate, not only when the caller uses the
-    /// predicate variable as a key to cache results. However, to avoid tight coupling between passes, we make the deduplication
-    /// conditional on whether the caller wants the predicate to be taken into account or not.
-    pub(crate) fn can_be_deduplicated(
-        &self,
-        function: &Function,
-        deduplicate_with_predicate: bool,
-    ) -> bool {
-        use Instruction::*;
-
-        match self {
-            // These either have side-effects or interact with memory
-            EnableSideEffectsIf { .. }
-            | Allocate
-            | Load { .. }
-            | Store { .. }
-            | IncrementRc { .. }
-            | DecrementRc { .. } => false,
-
-            Call { func, .. } => match function.dfg[*func] {
-                Value::Intrinsic(intrinsic) => {
-                    intrinsic.can_be_deduplicated(deduplicate_with_predicate)
-                }
-                Value::Function(id) => match function.dfg.purity_of(id) {
-                    Some(Purity::Pure) => true,
-                    Some(Purity::PureWithPredicate) => deduplicate_with_predicate,
-                    Some(Purity::Impure) => false,
-                    None => false,
-                },
-                _ => false,
-            },
-
-            // We can deduplicate these instructions if we know the predicate is also the same.
-            Constrain(..) | ConstrainNotEqual(..) | RangeCheck { .. } => deduplicate_with_predicate,
-
-            // Noop instructions can always be deduplicated, although they're more likely to be
-            // removed entirely.
-            Noop => true,
-
-            // Cast instructions can always be deduplicated
-            Cast(_, _) => true,
-
-            // Arrays can be mutated in unconstrained code so code that handles this case must
-            // take care to track whether the array was possibly mutated or not before
-            // deduplicating. Since we don't know if the containing pass checks for this, we
-            // can only assume these are safe to deduplicate in constrained code.
-            MakeArray { .. } => function.runtime().is_acir(),
-
-            // These can have different behavior depending on the EnableSideEffectsIf context.
-            // Replacing them with a similar instruction potentially enables replacing an instruction
-            // with one that was disabled. See
-            // https://github.com/noir-lang/noir/pull/4716#issuecomment-2047846328.
-            Binary(_)
-            | Not(_)
-            | Truncate { .. }
-            | IfElse { .. }
-            | ArrayGet { .. }
-            | ArraySet { .. } => {
-                deduplicate_with_predicate || !self.requires_acir_gen_predicate(&function.dfg)
-            }
-        }
-    }
-
-    /// Indicates if the instruction can be safely hoisted out of a loop.
-    /// If `hoist_with_predicate` is set, we assume we're hoisting the instruction
-    /// and its predicate, rather than just the instruction. Setting this means instructions that
-    /// rely on predicates can be hoisted as well.
-    ///
-    /// Certain instructions can be hoisted because they implicitly depend on a predicate.
-    /// However, to avoid tight coupling between passes, we make the hoisting
-    /// conditional on whether the caller wants the predicate to be taken into account or not.
-    ///
-    /// This differs from `can_be_deduplicated` as that method assumes there is a matching instruction
-    /// with the same inputs. Hoisting is for lone instructions, meaning a mislabeled hoist could cause
-    /// unexpected failures if the instruction was never meant to be executed.
-    pub(crate) fn can_be_hoisted(&self, function: &Function, hoist_with_predicate: bool) -> bool {
-        use Instruction::*;
-
-        match self {
-            // These either have side-effects or interact with memory
-            EnableSideEffectsIf { .. }
-            | Allocate
-            | Load { .. }
-            | Store { .. }
-            | IncrementRc { .. }
-            | DecrementRc { .. } => false,
-
-            Call { func, .. } => match function.dfg[*func] {
-                Value::Intrinsic(intrinsic) => intrinsic.can_be_deduplicated(false),
-                Value::Function(id) => match function.dfg.purity_of(id) {
-                    Some(Purity::Pure) => true,
-                    Some(Purity::PureWithPredicate) => false,
-                    Some(Purity::Impure) => false,
-                    None => false,
-                },
-                _ => false,
-            },
-
-            // We cannot hoist these instructions, even if we know the predicate is the same.
-            // This is because an loop with dynamic bounds may never execute its loop body.
-            // If the instruction were to trigger a failure, our program may fail inadvertently.
-            // If we know a loop's upper bound is greater than its lower bound we can hoist these instructions,
-            // but we do not want to assume that the caller of this method has accounted
-            // for this case. Thus, we block hoisting on these instructions.
-            Constrain(..) | ConstrainNotEqual(..) | RangeCheck { .. } => false,
-
-            // Noop instructions can always be hoisted, although they're more likely to be
-            // removed entirely.
-            Noop => true,
-
-            // Cast instructions can always be hoisted
-            Cast(_, _) => true,
-
-            // Arrays can be mutated in unconstrained code so code that handles this case must
-            // take care to track whether the array was possibly mutated or not before
-            // hoisted. Since we don't know if the containing pass checks for this, we
-            // can only assume these are safe to hoist in constrained code.
-            MakeArray { .. } => function.runtime().is_acir(),
-
-            // These can have different behavior depending on the predicate.
-            Binary(_)
-            | Not(_)
-            | Truncate { .. }
-            | IfElse { .. }
-            | ArrayGet { .. }
-            | ArraySet { .. } => {
-                hoist_with_predicate || !self.requires_acir_gen_predicate(&function.dfg)
-            }
-        }
-    }
-
-    pub(crate) fn can_eliminate_if_unused(&self, function: &Function, flattened: bool) -> bool {
-        use Instruction::*;
-        match self {
-            Binary(binary) => {
-                if matches!(binary.operator, BinaryOp::Div | BinaryOp::Mod) {
-                    if let Some(rhs) = function.dfg.get_numeric_constant(binary.rhs) {
-                        rhs != FieldElement::zero()
-                    } else {
-                        false
-                    }
-                } else {
-                    true
-                }
-            }
-
-            Cast(_, _)
-            | Not(_)
-            | Truncate { .. }
-            | Allocate
-            | Load { .. }
-            | ArrayGet { .. }
-            | IfElse { .. }
-            | ArraySet { .. }
-            | Noop
-            | MakeArray { .. } => true,
-
-            // Store instructions must be removed by DIE in acir code, any load
-            // instructions should already be unused by that point.
-            //
-            // Note that this check assumes that it is being performed after the flattening
-            // pass and after the last mem2reg pass. This is currently the case for the DIE
-            // pass where this check is done, but does mean that we cannot perform mem2reg
-            // after the DIE pass.
-            Store { .. } => {
-                flattened && function.runtime().is_acir() && function.reachable_blocks().len() == 1
-            }
-
-            Constrain(..)
-            | ConstrainNotEqual(..)
-            | EnableSideEffectsIf { .. }
-            | IncrementRc { .. }
-            | DecrementRc { .. }
-            | RangeCheck { .. } => false,
-
-            // Some `Intrinsic`s have side effects so we must check what kind of `Call` this is.
-            Call { func, .. } => match function.dfg[*func] {
-                // Explicitly allows removal of unused ec operations, even if they can fail
-                Value::Intrinsic(Intrinsic::BlackBox(BlackBoxFunc::MultiScalarMul))
-                | Value::Intrinsic(Intrinsic::BlackBox(BlackBoxFunc::EmbeddedCurveAdd)) => true,
-
-                Value::Intrinsic(intrinsic) => !intrinsic.has_side_effects(),
-
-                // All foreign functions are treated as having side effects.
-                // This is because they can be used to pass information
-                // from the ACVM to the external world during execution.
-                Value::ForeignFunction(_) => false,
-
-                // We must assume that functions contain a side effect as we cannot inspect more deeply.
-                Value::Function(_) => false,
-
-                _ => false,
-            },
-        }
     }
 
     /// If true the instruction will depend on `enable_side_effects` context during acir-gen.
