+++ compiler/noirc_evaluator/src/ssa/opt/loop_invariant.rs
@@ -60,8 +60,9 @@ use crate::ssa::{
         },
         post_order::PostOrder,
         types::Type,
-        value::ValueId,
+        value::{Value, ValueId},
     },
+    opt::pure::Purity,
 };
 
 use super::unrolling::{Loop, Loops};
@@ -317,9 +318,9 @@ impl<'f> LoopInvariantContext<'f> {
                 !self.defined_in_loop.contains(&value) || self.loop_invariants.contains(&value);
         });
 
-        let can_be_hoisted = instruction.can_be_hoisted(self.inserter.function, false)
+        let can_be_hoisted = can_be_hoisted(&instruction, self.inserter.function, false)
             || matches!(instruction, MakeArray { .. })
-            || (instruction.can_be_hoisted(self.inserter.function, true)
+            || (can_be_hoisted(&instruction, self.inserter.function, true)
                 && !self.current_block_control_dependent)
             || self.can_be_hoisted_from_loop_bounds(&instruction);
 
@@ -472,6 +473,78 @@ impl<'f> LoopInvariantContext<'f> {
                 self.inserter.push_instruction(instruction_id, block);
             }
             self.inserter.map_terminator_in_place(block);
+        }
+    }
+}
+
+/// Indicates if the instruction can be safely hoisted out of a loop.
+/// If `hoist_with_predicate` is set, we assume we're hoisting the instruction
+/// and its predicate, rather than just the instruction. Setting this means instructions that
+/// rely on predicates can be hoisted as well.
+///
+/// Certain instructions can be hoisted because they implicitly depend on a predicate.
+/// However, to avoid tight coupling between passes, we make the hoisting
+/// conditional on whether the caller wants the predicate to be taken into account or not.
+///
+/// This differs from `can_be_deduplicated` as that method assumes there is a matching instruction
+/// with the same inputs. Hoisting is for lone instructions, meaning a mislabeled hoist could cause
+/// unexpected failures if the instruction was never meant to be executed.
+pub(crate) fn can_be_hoisted(
+    instruction: &Instruction,
+    function: &Function,
+    hoist_with_predicate: bool,
+) -> bool {
+    use Instruction::*;
+
+    match instruction {
+        // These either have side-effects or interact with memory
+        EnableSideEffectsIf { .. }
+        | Allocate
+        | Load { .. }
+        | Store { .. }
+        | IncrementRc { .. }
+        | DecrementRc { .. } => false,
+
+        Call { func, .. } => match function.dfg[*func] {
+            Value::Intrinsic(intrinsic) => intrinsic.can_be_deduplicated(false),
+            Value::Function(id) => match function.dfg.purity_of(id) {
+                Some(Purity::Pure) => true,
+                Some(Purity::PureWithPredicate) => false,
+                Some(Purity::Impure) => false,
+                None => false,
+            },
+            _ => false,
+        },
+
+        // We cannot hoist these instructions, even if we know the predicate is the same.
+        // This is because an loop with dynamic bounds may never execute its loop body.
+        // If the instruction were to trigger a failure, our program may fail inadvertently.
+        // If we know a loop's upper bound is greater than its lower bound we can hoist these instructions,
+        // but we do not want to assume that the caller of this method has accounted
+        // for this case. Thus, we block hoisting on these instructions.
+        Constrain(..) | ConstrainNotEqual(..) | RangeCheck { .. } => false,
+
+        // Noop instructions can always be hoisted, although they're more likely to be
+        // removed entirely.
+        Noop => true,
+
+        // Cast instructions can always be hoisted
+        Cast(_, _) => true,
+
+        // Arrays can be mutated in unconstrained code so code that handles this case must
+        // take care to track whether the array was possibly mutated or not before
+        // hoisted. Since we don't know if the containing pass checks for this, we
+        // can only assume these are safe to hoist in constrained code.
+        MakeArray { .. } => function.runtime().is_acir(),
+
+        // These can have different behavior depending on the predicate.
+        Binary(_)
+        | Not(_)
+        | Truncate { .. }
+        | IfElse { .. }
+        | ArrayGet { .. }
+        | ArraySet { .. } => {
+            hoist_with_predicate || !instruction.requires_acir_gen_predicate(&function.dfg)
         }
     }
 }
